backend/requirements.txt
Django==2.1.7
django-rest-swagger==2.2.0
djangorestframework==3.9.2
psycopg2-binary==2.7.7
pytz==2018.9
django-filter==2.1.0
factory_boy==2.11.1
django-cors-headers==2.5.2